requirements.txt
streamlit
langchain
openai
faiss-cpu
pdfplumber
python-dotenv
sentence-transformers
